3/1/PBZ/2/requirements.txt
fastapi
uvicorn
psycopg2-binary
streamlit
pandas
requests
